topic: bug fix type: minor module: commons description: Use BigInteger instead of Integer.

+++ src/main/java/org/deegree/filter/function/other/IDiv.java
@@ -37,6 +37,7 @@ package org.deegree.filter.function.other;
 
 import static org.deegree.filter.function.other.IMod.extractValues;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.deegree.commons.tom.TypedObjectNode;
@@ -69,7 +70,7 @@ public class IDiv extends Function {
                             throws FilterEvaluationException {
         Pair<Integer, Integer> p = extractValues( getParams()[0], getParams()[1], f );
 
-        return new TypedObjectNode[] { new PrimitiveValue( p.first / p.second ) };
+        return new TypedObjectNode[] { new PrimitiveValue( BigInteger.valueOf( p.first / p.second ) ) };
     }
 
 }
